docker management: correct how we determine whether a container is running
We use `Running()` to determine if a container is running before we add it to the site config on server startup & stop language servers on shutdown.

But our `Running` implementation was flawed: it considered `StatusUnhealthy` to be 'running', which it is clearly not as it includes:

- Container not running
- Contained is paused by Docker
- Container was OOM killed.
- Container is dead (e.g. killed by Docker).

I think the reason this may have been done like this initially was because of OOM killed -- the container is _unhealthy_ but it is still running (i.e. Docker will start it again). In this case, we will now show a flicker (OOM killed -> "not running" -> "running" rather than OOM killed -> "running (still)") which makes more sense from a UX perspective. Note, however, that we do not use this `Running` method in the UI so it has no user-facing change.

Fixes #12559