docker management: do not report `docker rm`d language servers as healthy
Prior to this change language servers in the `langserver.StatusNone` state
were reported as `healthy`. This is not correct because the container not
existing (having no status) is an indicator that the container is unhealthy.

However, this commit introduces a new issue: GraphQL language server
enablement is asynchronous and thus as soon as you click "Enable" the
language server is "Unhealthy" in the UI for a few seconds until it is
actually `docker run` on the backend. This will be fixed in the next commit.

Fixes #10896