managed docker: do not display 'Unhealthy' when starting language servers
In the last commit, an issue was introduced as I previously described:

> However, this commit introduces a new issue: GraphQL language server
> enablement is asynchronous and thus as soon as you click "Enable" the
> language server is "Unhealthy" in the UI for a few seconds until it is
> actually `docker run` on the backend. This will be fixed in the next commit.

This commit fixes that issue by making it such that GraphQL language server
enablement does not return until the language server has actually been
`docker run` and the internal container info cache has been updated.

This is a bit complex change for what it is, but there isn't a better
option for right now. Perhaps in the future we can make some of this code
nicer with a few refactors -- but overall this is an improvement for now.