cmd/frontend: make it easier to test "no docker socket" mode (set `DEBUG_NO_DOCKER_SOCKET=t`)
This mode needs to be supported and, as such, tested occasionally. I used to
just make this change directly in the codebase, since it is trivial, but for
others who don't know which code to modify they would have to spin up a dev
server which is not quick. This is nicer.