graphqlbackend: langservers: refactor code to make Info requirement more obvious
Prior to this change, we relied on `infoErr == nil` to display a nice UI when
Docker management was not available. It is easy when looking at this code to
think that it only considers Server and Data Center, while neglecting the
"Server without managed Docker" case. It'd be easy for someone to introduce
`if conf.IsServer(conf.DeployType()) { return infoErr }` and think it is correct
but actually break the "Server without managed Docker" case by accident.

I've refactored this code here to make it more obvious that we must also support
that case and hopefully avoid any potential future regression.