Add sub-field indexes for global_dep (#12597)
* Add sub-field indexes for global_dep

An index on a json field can't be used to refer to specific
subfields, but that doesn't mean you can't make an index on
those fields.

Partially relevant:
https://www.postgresql.org/docs/9.4/static/datatype-json.html

In fact, because we want the actual text of the fields, rather
than json objects, we don't want to use a GIN index, and we
want to use a `->>` operator, but doing that results in EXPLAIN
of a query showing checking a hash bitmap, which should
be significantly faster.

Backwards compatibility: Not an issue, indexes only affect
performance. (There are probably exceptions, but they don't
seem to apply here.)

* Use collation to perform MIRACLES.

So it turns out that the LIKE operation still kills performance.
There's an obvious hack: don't use LIKE. Use strict ordering
comparisons.  Unfortunately, the default collation for PostgreSQL doesn't
let that obvious hack work, because it's not really ASCII order.

But you can work around this by qualifying values with
	`COLLATE "C"`

Unfortunately, `COLLATE` is an operator, which means that if you want
to do that, you do conversions on all of those indexed values,
which makes performance worse again.

But wait! The index can be on *any* expression, pretty much. So,
we change the index to be, not on `dep_data ->> 'package'`, but on
`dep_data ->> 'package' COLLATE "C"`. Then we modify the select
statement to match this.

Insane? Yes.

But on a 3,663-row table, the original code ran in a bit over a
millisecond, and the indexed code with LIKE still ran in a bit
over a millisecond, and this sucker runs in less than 50
microseconds.

* Use C collation for index to compare with boring string

I know this looks stupid, but it seriously makes a factor of five
or so difference in runtime.