fix oversight in migration (DANGER WILL ROBINSON) (#12641)
The query actually merged doesn't use COLLATE "C" for the depth query.
Drop that from the migration. Also fix names and add a missing comment.

If you are having problems because of this, drop the previous
indexes (idx_globaldep_package, idx_globaldep_depth), and run
the commands from 1528395539_.up.sql manually, and possibly
fix the migration table. Sorry. We were in a hurry and didn't
notice that a thing was still in my local tree but not up on
github.

We're doing this by directly overwriting an existing migration because
the migration in question takes so long to run on prod (with a 17M
row table) that it triggers the readiness-check killer and leaves
the database flagged unmigrated, and prevents frontend from launching
successfully.