cmd/frontend: stop langservers when they are removed from the site config (not disabled)
Prior to this change we did not stop language servers when they were _not in_
the site config. We only did it when they were `"disabled": true` in the site
config (which is surprising behavior). This has apparently been a bug for a
very long time that went unnoticed.

After this change, if language servers are disabled OR not in the site config,
we attempt to stop them. This inherently prevents them from e.g. being added
back to the site config as we detect them running.

Fixes #12676