drop a bad index (#12642)
It turns out that having an index on dep_data->>'depth' *hurts*
performance even when queries are hitting it a lot.

What's happening: Say we do a query looking for repos which
depend on 'github.com/sourcegraph' (note that this is not an
actual package name, but it generates more hits than most package
names do).

If there's no index on depth, we use the index on package,
find ~200 entries, and scan those for depth. This takes a
couple of milliseconds.

If there's an index on depth, we do both indexes in parallel
and bitwise-and their results. Because about 1/3 of repos have
depth = 0, this takes about 500ms.

(This is the sort of thing that might be better with later
postgres, but I'm not about to go investigating that right now.)