Allow per-process environment variables. (#12623)
* Allow per-process environment variables.

I keep wanting to run with debugging on repo-updater, and setting
it manually in Procfile means I have a changed procfile and git
fusses at me, so what if we supported per-process overrides for
environment variables? This is reasonably cheap and also exposes
an env.MyName which other code could use sometimes. Internally,
it uses the name with hyphens smashed to underscores so for
instance, you could use `SRC_LOG_LEVEL_repo_updater=dbug` to
change the log level for just one process.

Also convert tracer to use `env.MyName` as the default, so if we want to change how we express/pick those names some day, it'll be unified.