discussions: properly jump to lines when clicking discussion links
Previously, links to code discussions that did not contain the actual line
in the URL hash would never jump to the line the code discussion was created
on. Now we perform a redirect in this situation, e.g.:

```
http://localhost:3080/github.com/gorilla/mux/-/blob/mux.go#tab=discussions&threadID=1&commentID=1
http://localhost:3080/github.com/gorilla/mux/-/blob/mux.go#L18:29&tab=discussions&threadID=1&commentID=1
```

We also now only ever generate the links without the line number in the UI,
as this is the 'canonical' form. In the future, the line number may be
different depending on discussion reattachment logic for example.

Fixes #12390