discussions: do not render Markdown by default when visiting discussion links
Prior to this change, if you visited a code discussion link that was to a Markdown
or other file that we have a rich HTML view for, it would not open the discussion.
This is because we only have code discussions for the actual code views, currently,
not for the rich HTML view (which is probably not worthwhile right now).

To fix this, I've made it so that such links always go to the code view by default.

Fixes #12492