replace directories & files on tree pages with discussions
When (and only when) discussions is enabled, we now replace the directories and
files section of the repo tree pages (e.g. https://sourcegraph.sgdev.org/github.com/sourcegraph/sourcegraph)
with discussions instead.

I chose to replace the directories & files because:

1. With directories, files, commits, and discussions on the page it looks far too cluttered.
2. My assumption is that *most* people are actually using the file tree sidebar to navigate.
3. We can change in the future based on internal feedback, depending on what others here think.

The new discussions section shows a list of discussion threads made in any file
below the folder you are looking at.