fix(cxp): send didOpen before definition requests
Previously, CXP didOpen notifications needed to wait for the React <Blob> component to be rendered (and for the render/update cycle to fully finish). But definition requests (to load the definition for the position specified in the URL) would be fired by the <Blob> component's React lifecycle methods without waiting for the full render+update cycle. This meant that definition requests were sometimes sent before didOpen requests.

Known issues:

- Now, no didClose notification is sent when you navigate away from the file. This is OK for now.
- We may need to move the subscription (the newly added one in this commit, `// Update the CXP state to reflect the current file.`) higher up in the constructor to make sure the didOpen notif gets sent before the definition request.