Backoff scheduler in face of errors (#12799)
* Backoff scheduler in face of errors

One large customer reports a large density of repo-updater scheduled
attempts being failures. This is possibly because our fallback if we
didn't get accurate information is to assume we should check again
soon -- which is wrong if checks are failing unconditionally.

Check for errors, and on error, backoff exponentially (duration * 2)
up to a maximum of a day.

Also, stop reporting "recently fetched, skipping" as an error, because
it's really not an error, and we don't want it to trigger backoff.