Draft proposal for making scopes ACL-like
This is possibly unnecessary or premature, but I really don't
like making up magic cookies without some kind of schema in mind
that they fit in. This draft design would give us a framework
for thinking about scopes which I think offers at least reasonable
core functionality, and is probably easy to migrate later to
a more structured system if we need to, but which also admits
the quick and dirty implementation of a token scope which
only allows a webhook to request repo updates.