Fix detection for "wild" repos. (#12851)
A customer requested that we filter out repository patterns, which
gitolite reports in the same list as repositories, without any
reliable distinguishing characteristics. It turns out that we
already have a filter for this, our filter just disagrees with
gitolite.

Gitolite's internal configuration for this feature works by
having two regexes, one which matches valid repository names,
and another which matches valid repository patterns. The two
sets are nearly identical, the difference being that a
repository pattern may begin with `[`, and repository patterns
may contain a number of regex characters after that first
character. Since any string containing those characters would
be rejected as a literal repository name, finding any of those
characters in a string tells us it must be a pattern. So we
just check for those.