fix: eliminate data races in test code
There was a data race on reads of MockGetData in conf.Get, and writes
to MockGetData in various test cases. This was discovered because
the CI system ran slowly enough that something's conf.Watch() triggered
during a test, which isn't a very common outcome (configuration
doesn't normally change during testing).

A similar data race existed in gitserver from the possibility of
`originMaps.setup` being called during tests, resulting in racey writes
to the internals of originMaps. In addition to the race itself, this has the
problem that those writes could break the tests. We now suppress
configuration change processing while the tests are running.

Fixes #12861.