feat: preserve null and errors in parsed config
Previously, the API was unclear about whether it propagated errors that occurred while fetching and parsing configuration. Now, there are `XyzOrError` types that make it explicit.

BREAKING CHANGE: Users of this package need to use the new `XyzOrError` types and be more explicit when passing in configuration values that might be `ErrorLike`.