feat: remove all platforms other than JS bundle
Going forward, only CXP extensions that are compiled to JavaScript and that can run on the client are supported, for several reasons (latency, privacy, security, etc.).

BREAKING CHANGE: The extension manifest `platform` property was removed, and a new property `url` (which corresponds to the old `url` property in the `BundleTarget`) was added.