refactor: remove `Windows#all`
Now that the `CXP#activeWindow` accessor is at the top level, we get a simplicity win by removing this field so that `Windows` has no non-method fields. Any extension author needing the list of *all* windows is probably comfortable with using Observables.

BREAKING CHANGE: Subscribe to the `CXP#windows` Observable to obtain the list of all windows.