fix: prevent creation of extensions with the same name
fix #12306

A bug in the PostgreSQL unique index allowed the creation of multiple extensions with the same publisher and name, such as two extensions named `alice/foo` and `alice/foo`. This commit migrates the DB so that the unique index coalesces `NULL` values to `0`, which is necessary because PostgreSQL considers `NULL` values equal to nothing, not even other `NULL` values.