feat: use a single namespace for usernames and org names (#12982)
fix #12068

Enforces that no user may be created with a name taken by an existing org, and vice versa.

In the past, it was possible to create a user and organization with the same name. This creates confusion in new features that treat users and organizations as sharing a single namespace (or will do so soon), such as extension IDs (where the publisher can be either a user or organization, such as `alice/myextension`) and @-mentions in code discussions.