docker management: avoid killing langservers that provide multiple languages, unless all disabled
When a language server is enabled/disabled in the UI we just change the site
config and rely on the site config watcher to reflect the changes in the actual
system Docker state.

When `javascript` and `typescript` languages are `enabled` in the site config,
and one becomes `disabled`, we previously would try to stop the container of
the `disabled` one. This is fine for most languages, but `javascript` and
`typescript` are provided by the same language server and so inherently we
would accidentally kill the language server of the other.

This regression was introduced in 5c4c04f.

This is now fixed by ensuring that before we stop the Docker container it is
no longer in use by any language it provides. e.g. we only stop the container
if both `javascript` AND `typescript` are disabled, not just one.

A new `ForceStop` method is added which ignores this restriction, and is used
on e.g. site shutdown where such behavior is not desired.

Fixes #12998