feat: remove unused ConfigurationCascade.defaults (#13050)
This was always an empty JSON object and was never used. If we need to introduce defaults at a lower precedence than global settings, we should insert them as a ConfigurationSubject in ConfigurationCascade.subjects instead of special-casing them as another field on ConfigurationCascade.

BREAKING CHANGE: This removes the GraphQL API ConfigurationCascade.defaults field, which was not used anywhere.