discussions: fetch updated thread every 2.5s, not constantly (thousands or requests)
Prior to this change, whenever the updated thread request would finish we would
update the component state which in turn fired `componentDidUpdate` thus firing
`this.componentUpdates`. This is intended.

However, because `distinctUntilChanged` was comparing `a.threadID !== b.threadID`,
we never prevented identical `threadID` values from continuing the observable chain.
i.e., the comparator function passed to `distinctUntilChanged` is expected to check
for equality, NOT inequality and as such it was roughly useless before this change.

Fixes #13077