mv docs/analytics.md enterprise/docs/
I am moving this to our private codebase as this is primarily going to be
useful for only us internally, and even if we do use such analytics in our
open source builds, we need to be very careful about how we communicate it due
to the "analytics & tracking == EVIL privacy invasion!!11" mindset of the
public. i.e., this document does more harm than good for our open source
contributors / users of tomorrow.

cc @dadlerj