refactor: reorganize the project to fit its current state and be easier to understand
This is a reorganization of the code base. This code base was originally just a Chrome extension. This is where the original architecture started and at the time it was clean. As we expanded the code base to be much more than just the Chrome extension, we neglected to reorganize to keep things clean. This is a long overdue start to that!

The overall goal of this PR is to make the project easier to understand and navigate when you first look at the code and onward.

Resolves #54.