Make UserAccountArea routes customizable (#13135)
#13090 

These are a bit more complex, so this introduces a new field `condition` on configuration objects, which gets called to determine if the route should be handled. I am extracting this into generic reusable interfaces, where the context that is passed along to `condition` and `render` is passed as a type parameter.