refactor: remove now-unneeded direct LSP code
Now all LSP queries can go through the sourcegraph-langserver-http-based extensions (which are added automatically based on the site config `langservers` value). These extensions do the same thing as the code that this commit removes. It is better to have this logic live in extensions because it lets us build more custom language features and makes the interface unified for all sources of hovers/definitions/etc. (LSP and non-LSP).

The sourcegraph-langserver-http-based code path has been in use for 2 weeks on Sourcegraph.com without major flaws. It was not shipped in 2.11, but by 2.12 it will have been so thoroughly tested that we do not need to keep it in for another major release.