List all contributors to the app
Update the mailmap file to list/remove all duplicate email addresses
referring to the same user.

Are there any people who might want to have a "public" email address
listed, instead of the email address used in the commit history? We
will have a chance to remove/change the aliases when we flatten the
history here.

Fixes #13314.