repo-updater: add option to exclude personal BitBucket repositories (#13226)
* repo-updater: add option to exclude personal BitBucket repositories

This PR adds an option to exclude personal BitBucket repositories for users who
do not wish to create a Sourcegraph bot accoutn on their BitBucket instance,
for example.

When someone adds Bitbucket Server integration to their Sourcegraph instance,
the access token they provide may be e.g. created by an admin account. In this
situation the admin account has access to view and clone *ALL* personal
repositories on Bitbucket. These repositories are typically hidden from e.g.
other users:

> By default, personal repositories are not visible to other Bitbucket Server users (unless they are a Bitbucket Server system administrator). However, you can:
>
> - use repository permissions to open up access to other Bitbucket Server users and groups, for collaboration or review.
> - allow public access (read-only) to your project, for anonymous users.

While the suggested way to fix this is to create a Sourcegraph bot account on
Bitbucket and only give Sourcegraph access to that, this provides a secondary
escape hatch for users in the case that they do not wish to do that.

The recommended path for all users is still to make use of a bot account, and
the documentation is updated to reflect this already.

Helps sourcegraph/issues#142

* CHANGELOG: update to reflect Bitbucket personal repositories change

* document default