Check that all authors have been added to AUTHORS
github.com/kevinburke/differ will check Git status before and after
running a given command. Use it to run the `generate.sh` script and
also to check whether the AUTHORS file is up to date.

This technically changes the behavior of the generate check - the
previous script would preserve the contents of an existing git diff
and only error if that changed - but the only thing that runs the
checks is the build pipeline and I think it's appropriate to ensure
that there are no stray files before or after, instead of just
ensuring the same stray files before are present afterwards.

Fix template-inlines.sh to check for templates in the appropriate
place - the script currently warns that app/templates does not exist.
Also, you need to separate variable declaration and subscript running
in order to get the subscript to exit with an error, if/when there is
an error.