Call Output() so ExitErr.Stderr gets populated
One branch of this code would call cmd.Run(), which does not
populate ExitErr.Stderr if it exits with a non-zero return value. This
meant that a confusing error would be logged to the screen, which
I spotted when running the quickstart from the website.

Update the code so Output() is always used to run the code and Stderr
is always populated.

Also file https://github.com/dominikh/go-tools/issues/343 to see if
this can trigger a lint check.