Fix tar unpack RCE on Alpine image
The base Sourcegraph image was being built with Alpine 3.6 which is
vulnerable to an RCE as described here:
https://justi.cz/security/2018/09/13/alpine-apk-rce.html

Fix the issue by updating to Alpine 3.8. Alpine Docker does not
release patch version tags, so we'll need/want to double check that we
have the latest 3.8 version included on the machine that builds the
server image.

We also probably want to release/upload new versions of the Docker
image to customers.