pkg/processrestart: clean up process restart logic
As written, the error handling will never trigger, because we assign
`runErr` as the response from cmd.Run(), but then check `err` to
determine whether to run error handling. As a result, `err` is never
non-nil and we never get a chance to assign `runErr` to it.

Simplify the logic by using an errgroup and immediately returning
when we get any error, which should cancel the other subprocesses
immediately.