feat: show Sourcegraph license to site admins, extend its API in enterprise
Previously, the "Sourcegraph license" shown on the site admin page really just showed what premium features were in use and how many users were signed up. There was no license key.

Now, a license key is provided in site config (`sourcegraphLicense`). Customers obtain this by paying us. When the license key is supplied, the enterprise build injects a "License" site admin page and a component into the SiteAdminOverviewPage that display the license information (plan, max user count, and expiration date). If the license is not provided, it falls back to showing "Free" with unlimited users and no expiration date. The exact language here will be changed before release.

There is no enforcement of the license yet in this commit. That will come in a subsequent commit.