cmd/server, dev/src-prof-services.json: sync debug ports
The debug port for frontend conflicts with the most commonly used port
for godoc, a service that is common for Go developers to run locally.
Update the port number to 6063 from 6060 to avoid conflicts.

Add xlang-go to the list of services in cmd/server/shared/globals.go,
to keep it in sync with src-prof-services.json.

Fixes #13402.