README.md: remove icons
I don't think they add much in particular and they make us look bad if
the build is failing. It's easy to see whether the build is failing
because the build status is present on any PR/tip commit on master,
and you can tell we use prettier because the build will fail if your
formatting doesn't match it.