all: Remove dependency on go-langserver
We use go-langserver for the structs it exports related to LSP. However,
go-langserver is versioned by the releases of the command, not as a library.
This lead to complications in gomod. Additionally what we used was very minor
but used in many places in our codebase. As such the used pkgs are copied into
the sourcegraph repo.