feat: support true-up user licensing model with true-up license tag
To create a license that allows truing up, give it a `true-up` tag during license generation.

If a license has the true-up tag, then new users may sign up or be created beyond the license's user count. A server log will be printed saying that the user will be charged in the future.

This alone is not sufficient notice for admins. If we create a true-up license, we need to communicate directly and explicitly to the admin about how it works.