support allowing only explicitly specified remote extensions
This was moved from OSS because it is Enterprise functionality and we now have the ability to perform these feature checks easily in this repo.

The new `extensions.allowRemoteExtensions` site config, if non-null, specifies the sole list of remote extension IDs allowed. This is used by site admins who don't want to let their users install untrusted extensions that are published to Sourcegraph.com.

This feature is only available on Sourcegraph Enterprise (not Core or Enterprise Starter).