fix where the private ext registry license check occurs
It needs to be on viewerPublishers because that is what populates the "new extension" form. Previously it was on publishers, which is just what's called when querying for the whole list of publishers (not actually used in the UI right now).

There is ALSO the check that occurs on the actual mutation to publish, which would've still been triggered and enforced the license. That remains unchanged. The reason why we also check the license here is so that the user finds out immediately on the "new extension" form page before they even spend time choosing a name, etc.