Combine marketing and product homepage (#156)
New marketing homepage for sourcegraph.com!

Router users visiting about.sourcegraph.com/ and sourcegraph.com/ to the new homepage based off the following conditions:

If a user is logged out -> / -> /start and gets the new homepage
If a user is logged in -> / -> /search and gets the normal search page

Users that visit either /start or /search will get the page without redirections

This is not implemented on non-sourcegraph.com instances.