Point thread-loader at Sourcegraph fork (#13586)
The thread-loader package starts several subprocesses. When you send
a SIGINT to thread-loader, the subprocesses terminate, but the parent
process hangs while waiting for them to send data. We were not
listening for the 'end' event in the parent process which meant that
the webpack processes would just hang forever.

This is a problem because this blocks the shutdown of goreman which
means we can have orphan processes and is, in general, a huge hassle
when developing the codebase locally.

Fix this by listening for an 'end' event, which means we can signal to
Webpack that the job terminated with an error, instead of not
terminating. This means that SIGINT will actually shut down the
process instead of leaving it hanging forever.

Fixes sourcegraph/sourcegraph#186.
Updates webpack-contrib/thread-loader#33.
Updates webpack-contrib/thread-loader#34.
Updates webpack-contrib/thread-loader#35.
Updates webpack-contrib/thread-loader#36.