use docsite markdown package instead of github_flavored_markdown
The github.com/sourcegraph/docsite/markdown package performs roughly the same formatting but uses a newer underlying blackmonday library. This commit removes external dependencies but otherwise is a noop. Removing the dependency on github_flavored_markdown is necessary for a future commit where we need the new version of blackmonday (and it can't coexist with an older version at the same import path).