feat: use access tokens for authentication (#245)
This PR changes request logic to use access tokens. This means that admins won't have to add their code hosts to the `corsOrigin` anymore.

This fixes `corsOrigin` errors because we are able to make requests to the API from the options menu. Any request from the options menu will create an access token. I landed on creating a token on each request rather than explicitly creating a token when entering and saving a new URL because multiple requests go out at different times and it was difficult to make this happen before other requests go out. The easiest way was to get to the root of each request. 

Without this architecture change/decision, some requests would fail with a 401 which could cause problems. You would have to click save and then refresh the page for everything to work properly.