add docs to this repo, show them at /help (#421)
* move docs/*.md -> doc/dev/*.md

* use docsite markdown package instead of github_flavored_markdown

The github.com/sourcegraph/docsite/markdown package performs roughly the same formatting but uses a newer underlying blackmonday library. This commit removes external dependencies but otherwise is a noop. Removing the dependency on github_flavored_markdown is necessary for a future commit where we need the new version of blackmonday (and it can't coexist with an older version at the same import path).

* remove HelpPopover, just link to /help

This regresses on showing extension-contributed help action items, but that is OK and will be fixed before the next release.

* serve docsite in UI at /help

Docs from the top-level doc/ dir are served at the /help path. Content assets (such as .png files referenced in Markdown files) are also served at that path.

* remove old legacy redirects

* WIP add/update doc content

* do not try to watch now-removed vendor dir

* go.sum

* Removed unused images