support a billing product minQuantity, show "$10/year total for the first N users"
If the billing product has a minQuantity metadata item with a numeric value, it is used as a minimum quantity. The price for it is also shown as the annual total for that quantity, instead of per-user/per-month values.

This achieves the "<= 10 users is $10, 11th+ user is $4/month" that we want, with minimal add'l user complexity.