refactor: rename React user prop to authenticatedUser
Previously, the `user` prop could be either:

1. The subject user (e.g., alice in the React component for /users/alice/settings)
2. The currently authenticated user

Given any instance of `this.props.user` or `props.user`, it was not easy to determine whether it referred to the subject user (1) or the currently authenticated user. In the UserArea, the name `authenticatedUser` was used for (2).

This makes it so that `authenticatedUser` is used everywhere to refer to the currently authenticated user (2). The only places that the `user` prop name is used are where it refers to the subject user (1).

There are no intended user-facing effects of this.