feat: show price of subscription to unauthed viewers
This gives a friendlier buying experience to unauthed viewers (vs. just dumping them on a sign-in page).

The new subscription page for authed viewers is unchanged.

For unauthed viewers, they can input user count and choose a plan to see the quoted price. To continue, they need to click a big "Create account or sign in to continue" button, which takes them to signup/sign-in. After signing up or signing in, they're taken back to the new subscription form (with their original values re-populated).