schema: Add search.index config to toggle indexed search
Additionally we update the DB layer to optionally filter repositories based on
the search.index setting. When indexed search lists repositories it will use
this filter. So if disabled no repositories will be indexed. When enabled all
enabled repositories will be indexed.

The default value for search.index is auto which matches current behaviour
where indexed search is enabled for k8s, but disabled for our docker image.