feat: remove client settings for Sourcegraph extensions
Client settings (eg settings stored in your browser extension) add complexity and are not worth supporting right now. This means that you must be authenticated to persist settings changes. We can mitigate most of the downsides of this and make it so unauthed users still can use most features of the default-enabled extensions by storing state in the live Web Worker process for unauthed users (such as whether a line decoration for code coverage should be shown).