fix: do not display code intelligence hovers on Sourcegraph Extension…
…line decorations

Prior to this change, hovering over text/links provided by a Sourcegraph Extension would trigger the display of the hover 'loading' indicator. On most files, it goes away quickly (just a flicker), but on some it displays an error:

![image](https://user-images.githubusercontent.com/3173176/45585161-caaff580-b894-11e8-8b3f-10347f2aa438.png)

After this change, we no longer attempt to display hovers etc. on the DOM elements which Sourcegraph Extensions provide.

Fixes sourcegraph/sourcegraph-extension-api#67