feat: user dropdown menu for account, settings, extensions, etc.
Previously, the top nav had too many links. They took up a lot of width and added clutter. Now, there is a user dropdown menu. It also includes a sign-out button, which is a fairly frequent user request.

If there is no authenticated user, it still shows a "Sign in" button and some of the links. Authenticated users no longer see "Saved searches", "Discussions", and the theme switcher (because it soon will persist to user settings, not localStorage).