feat: remove Explore page, make Repositories link to `repo:` search query
The Explore page was a duplicative list of repositories. The query box did not support regexps, the page was not well designed, and we want users to use `repo:` to find repositories anyway. So the Explore page is removed, and instead there is a Repositories link in the top navbar that links to a `repo:` query (which shows all repositories because there is no pattern specified).