remove Docker lang server management
Automatic management (start/stop/update/status/etc.) of language server Docker containers is being removed in favor of providing language support via Sourcegraph extensions. Many of those extensions will, in turn, rely on language servers for analysis, but that is beyond Sourcegraph's concern. In the future, however, we will add back a way for Sourcegraph extensions to ask the server to run a Docker image and make it available to the extension, to make that common case easy.

This commit DOES NOT update the docs to reflect the removal of Docker lang server management because those steps are still valid for the last release. Before releasing 3.0-preview, the docs need to be updated to reflect the new way to get language support.