refactor: allow options page to make token-authed fetches
The options page calls fetchCurrentUser to determine whether to show the user settings or client settings. This fetchCurrentUser request needs to be made with the same authentication (if any) as would be used on the content page, or else it would be confusing: the content page's extensions would read from user settings (using the token) but the options page would say the user is not authenticated (because it would not use the token).