feat: remove file tree
Reason: It is not the primary focus/draw of this extension, and we don't plan to make this available on other code hosts (so it leads to inconsistencies in product offerings among code hosts).

BREAKING CHANGE: The GitHub file tree is removed. Use [Octotree](https://chrome.google.com/webstore/detail/octotree/bkhaagjahfmjljalopjnoealnfndnagc) instead for this feature.