feat: remove code host search box integration
(The location bar search autocomplete feature remains.)

Reason: The experience was not great (it opened results in a new window) and it was disabled by default. It is not worth maintaining a feature that is disabled by default and that has no current plan for improvement.

BREAKING CHANGE: The "Open a new window with Sourcegraph search results when you perform a search on your code host." feature was removed. See https://docs.sourcegraph.com/integration/browser_search_engine for a better solution that is provided by the browser extension.