feat: remove CodeIntelStatusIndicator
This is the plug icon that showed the status of code intelligence. It is being removed because extensions will provide this indicator in the future in a custom, per-language/per-extension way. This indicator provides limited value because it will point users to the deprecated/old way of adding language support. It also does not work for Sourcegraph.com because the langserver-related GraphQL API fields have been removed.